src/requirements.txt
absl-py==0.5.0
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
branca==0.3.0
census==0.8.7
certifi==2018.11.29
cffi==1.11.4
chardet==3.0.4
citipy==0.0.5
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
conda==4.5.12
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.4.2
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
Django==2.1.5
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.13
Flask==1.0.2
Flask-Cors==3.0.3
Flask-PyMongo==2.1.0
Flask-SQLAlchemy==2.3.2
folium==0.5.0
future==0.16.0
gast==0.2.0
geojson==2.3.0
geoplotlib==0.3.2
gevent==1.2.2
glob2==0.6
gmaps==0.8.0
gmplot==1.2.0
gmpy2==2.0.8
graphviz==0.8.4
greenlet==0.4.12
grpcio==1.12.1
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
inflection==0.3.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.2
jupyterthemes==0.19.1
kdtree==0.16
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lesscpy==0.13.0
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mglearn==0.1.7
mistune==0.8.3
more-itertools==4.2.0
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
mysqlclient==1.3.13
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.5.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.5.0
numba==0.36.2
numexpr==2.6.4
numpy==1.14.4
numpydoc==0.7.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.4.10
openweathermapy==0.6.6
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.2.0
pipenv==2018.7.1
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
py4j==0.10.7
pyasn1==0.4.3
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pydotplus==2.0.2
pyflakes==1.6.0
pyglet==1.3.2
Pygments==2.2.0
pylint==1.8.2
pymongo==3.7.0
PyMySQL==0.9.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pyspark==2.3.2
pytest==3.3.2
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qhue==1.0.12
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
Quandl==3.3.0
requests==2.18.4
requests-oauthlib==1.0.0
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
selenium==3.14.0
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
splinter==0.8.0
spyder==3.2.6
SQLAlchemy==1.2.11
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
tweepy==3.6.0
tweetpy==0.1
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
vaderSentiment==3.2.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.2.0
zict==0.1.3
absl-py==0.5.0
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
branca==0.3.0
census==0.8.7
certifi==2018.11.29
cffi==1.11.4
chardet==3.0.4
citipy==0.0.5
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
conda==4.5.12
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.4.2
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
Django==2.1.5
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.13
Flask==1.0.2
Flask-Cors==3.0.3
Flask-PyMongo==2.1.0
Flask-SQLAlchemy==2.3.2
folium==0.5.0
future==0.16.0
gast==0.2.0
geojson==2.3.0
geoplotlib==0.3.2
gevent==1.2.2
glob2==0.6
gmaps==0.8.0
gmplot==1.2.0
gmpy2==2.0.8
graphviz==0.8.4
greenlet==0.4.12
grpcio==1.12.1
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
inflection==0.3.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.2
jupyterthemes==0.19.1
kdtree==0.16
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lesscpy==0.13.0
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mglearn==0.1.7
mistune==0.8.3
more-itertools==4.2.0
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
mysqlclient==1.3.13
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.5.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.5.0
numba==0.36.2
numexpr==2.6.4
numpy==1.14.4
numpydoc==0.7.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.4.10
openweathermapy==0.6.6
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.2.0
pipenv==2018.7.1
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
py4j==0.10.7
pyasn1==0.4.3
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pydotplus==2.0.2
pyflakes==1.6.0
pyglet==1.3.2
Pygments==2.2.0
pylint==1.8.2
pymongo==3.7.0
PyMySQL==0.9.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pyspark==2.3.2
pytest==3.3.2
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qhue==1.0.12
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
Quandl==3.3.0
requests==2.18.4
requests-oauthlib==1.0.0
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
selenium==3.14.0
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
splinter==0.8.0
spyder==3.2.6
SQLAlchemy==1.2.11
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
tweepy==3.6.0
tweetpy==0.1
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
vaderSentiment==3.2.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.2.0
zict==0.1.3
